Pre lane merge analysis
Before I can merge the lanes from sequencing together, I needed to check that the lanes were successful individually which meant processing them individually up to the markduplicates (2.1) point of the pipeline. Once running this and checking the duplication rate of all samples. I can then decide whether to merge the lanes or not. In this analysis, I will as duplication rates were low